# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QUANTEDGE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlists:
    get:
      operationId: getWatchlists
      tags: [watchlist]
      summary: List all watchlists
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watchlist"
    post:
      operationId: createWatchlist
      tags: [watchlist]
      summary: Create a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
  /watchlists/{id}:
    delete:
      operationId: deleteWatchlist
      tags: [watchlist]
      summary: Delete a watchlist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /watchlists/{id}/stocks:
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add stock to watchlist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistStockInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistStock"
  /watchlists/{id}/stocks/{symbol}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove stock from watchlist
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content

  # STOCKS
  /stocks/search:
    get:
      operationId: searchStocks
      tags: [stocks]
      summary: Search stocks
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
  /stocks/{symbol}:
    get:
      operationId: getStockDetail
      tags: [stocks]
      summary: Get stock research page data
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockDetail"
  /stocks/{symbol}/candles/{timeframe}:
    get:
      operationId: listStockCandles
      tags: [stocks]
      summary: Get OHLCV candle data
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
        - in: path
          name: timeframe
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /stocks/{symbol}/indicators/{timeframe}:
    get:
      operationId: listStockIndicators
      tags: [stocks]
      summary: Get technical indicators
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
        - in: path
          name: timeframe
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicalIndicators"

  # FORECASTS
...
components:
  schemas:
...
    Watchlist:
      type: object
      required: [id, name, stocks, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        stocks:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistStock"
        createdAt: { type: string }
    WatchlistInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    WatchlistStockInput:
      type: object
      required: [symbol]
      properties:
        symbol: { type: string }
    WatchlistStock:
      type: object
      required: [id, symbol, company, price, change, changePct, volume, forecastDirection, signal, confidence]
      properties:
        id: { type: integer }
        symbol: { type: string }
        company: { type: string }
        price: { type: number }
        change: { type: number }
        changePct: { type: number }
        volume: { type: number }
        forecastDirection: { type: string }
        signal: { type: string }
        confidence: { type: number }

    # STOCKS
    StockSearchResult:
      type: object
      required: [symbol, company, sector, exchange]
      properties:
        symbol: { type: string }
        company: { type: string }
        sector: { type: string }
        exchange: { type: string }
    StockDetail:
      type: object
      required: [symbol, company, sector, industry, description, marketCap, eps, pe, bookValue, dividendYield, roe, roa, debtToEquity, revenue, netProfit, operatingMargin, freeCashFlow, week52High, week52Low, beta, volatility, avgVolume, currentVolume, currentPrice, change, changePct]
      properties:
        symbol: { type: string }
        company: { type: string }
        sector: { type: string }
        industry: { type: string }
        description: { type: string }
        marketCap: { type: number }
        eps: { type: number }
        pe: { type: number }
        bookValue: { type: number }
        dividendYield: { type: number }
        roe: { type: number }
        roa: { type: number }
        debtToEquity: { type: number }
        revenue: { type: number }
        netProfit: { type: number }
        operatingMargin: { type: number }
        freeCashFlow: { type: number }
        week52High: { type: number }
        week52Low: { type: number }
        beta: { type: number }
        volatility: { type: number }
        avgVolume: { type: number }
        currentVolume: { type: number }
        currentPrice: { type: number }
        change: { type: number }
        changePct: { type: number }
...
    Candle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time: { type: string }
        open: { type: number }
        high: { type: number }
        low: { type: number }
        close: { type: number }
        volume: { type: number }
    TechnicalIndicators:
      type: object
      required: [sma20, sma50, ema20, rsi, macd, macdSignal, macdHistogram, bbUpper, bbMiddle, bbLower, vwap]
      properties:
        sma20: { type: number }
        sma50: { type: number }
        ema20: { type: number }
        rsi: { type: number }
        macd: { type: number }
        macdSignal: { type: number }
        macdHistogram: { type: number }
        bbUpper: { type: number }
        bbMiddle: { type: number }
        bbLower: { type: number }
        vwap: { type: number }

    # FORECASTS
...