# lib/api-zod/src/generated/types/technicalIndicators.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEDGE API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TechnicalIndicators {
  sma20: number;
  sma50: number;
  ema20: number;
  rsi: number;
  macd: number;
  macdSignal: number;
  macdHistogram: number;
  bbUpper: number;
  bbMiddle: number;
  bbLower: number;
  vwap: number;
}