# lib/api-zod/src/generated/types/watchlist.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEDGE API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistStock } from './watchlistStock';

export interface Watchlist {
  id: number;
  name: string;
  stocks: WatchlistStock[];
  createdAt: string;
}